Add cancel functionality to vesting contracts
This change adds the ability for an owner of vesting contracts to cancel the vesting schedule of
a token or ETH that is currently locked within it, returning unvested tokens back to the owner.

Vesting contracts retain the address of their deployer and read ownership from the factory that
deployed them, allowing one ownership transfer to effect all deployed vesting contracts.